Envio con post y usando json

+++ demo-sb2/src/main/java/co/edu/utp/misiontic/cesardiaz/demosb2/controller/dto/OperarRequest.java
@@ -5,15 +5,15 @@ package co.edu.utp.misiontic.cesardiaz.demosb2.controller.dto;
  */
 public class OperarRequest {
 
-    private Integer num1;
+    private Double num1;
     private String op;
-    private Integer num2;
+    private Double num2;
 
-    public Integer getNum1() {
+    public Double getNum1() {
         return num1;
     }
 
-    public void setNum1(Integer num1) {
+    public void setNum1(Double num1) {
         this.num1 = num1;
     }
 
@@ -25,11 +25,11 @@ public class OperarRequest {
         this.op = op;
     }
 
-    public Integer getNum2() {
+    public Double getNum2() {
         return num2;
     }
 
-    public void setNum2(Integer num2) {
+    public void setNum2(Double num2) {
         this.num2 = num2;
     }
